chat/requirements.txt
# FastAPI async chat application
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Socket.IO for real-time communication
python-socketio>=5.11.0

# Async SQLAlchemy
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0

# Async HTTP client
httpx>=0.26.0
aiohttp>=3.9.0

# Async SMTP
aiosmtplib>=3.0.0

# Pydantic for settings
pydantic>=2.5.0
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0

# Authentication
PyJWT>=2.8.0
werkzeug>=3.0.0
python-dotenv>=1.0.0

# Security
cryptography>=42.0.0
bleach>=6.1.0
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0

# PostgreSQL sync driver (for health checks)
psycopg2-binary>=2.9.9
